refs #117 Changed some minor things about how the linking to external pages is done.
